feat: restructure game components and implement pause functionality
