+++ package.json
@@ -11,21 +11,22 @@
   },
   "homepage": "https://github.com/harnyk/the-drift",
   "dependencies": {
+    "clsx": "^2.1.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "vercel": "^42.3.0"
   },
   "devDependencies": {
+    "@types/jest": "^29.5.2",
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
     "@vitejs/plugin-react": "^4.5.0",
     "autoprefixer": "^10.4.14",
+    "jest": "^29.6.0",
     "postcss": "^8.4.21",
     "tailwindcss": "^3.3.2",
-    "typescript": "^5.0.4",
-    "vite": "^4.3.0",
-    "jest": "^29.6.0",
     "ts-jest": "^29.1.1",
-    "@types/jest": "^29.5.2"
+    "typescript": "^5.0.4",
+    "vite": "^4.3.0"
   }
 }
