+++ src/game/Game.ts
@@ -1,23 +1,23 @@
-import { Context } from './engine/Context';
-import { FixedTimestepIntegrator } from './engine/FixedTimestepIntegrator';
-import { fromDeg } from './engine/fromDeg';
-import { CollisionDetector } from './engine/physics/CollisionDetector';
-import { Vec2D } from './engine/vec/Vec2D';
-import { Vec2DAverager } from './engine/Vec2DAverager';
-import { Viewport } from './engine/Viewport';
-import { World } from './engine/World';
-import { WorldRenderer } from './engine/WorldRenderer';
-import { Block } from './game/Block';
-import { Car } from './game/Car';
-import { KeyboardControl, KeyCodeWASD } from './game/controls/KeyboardControl';
-import { GameStateManager } from './game/GameStateManager';
-import { CompassRenderable } from './game/renderables/CompassRenderable';
-import { CurvedGrid } from './game/renderables/CurvedGrid';
-import { GameStateOverlayRenderable } from './game/renderables/GameStateOverlayRenderable';
-import { SpeedometerRenderable } from './game/renderables/SpeedometerRenderable';
-import { TerroristEyesRenderable } from './game/renderables/TerroristEyesRenderable';
-import { TerroristIndicatorRenderable } from './game/renderables/TerroristIndicatorRenderable';
-import { Terrorist } from './game/Terrorist';
+import { Context } from '../engine/Context';
+import { FixedTimestepIntegrator } from '../engine/FixedTimestepIntegrator';
+import { fromDeg } from '../engine/fromDeg';
+import { CollisionDetector } from '../engine/physics/CollisionDetector';
+import { Vec2D } from '../engine/vec/Vec2D';
+import { Vec2DAverager } from '../engine/Vec2DAverager';
+import { Viewport } from '../engine/Viewport';
+import { World } from '../engine/World';
+import { WorldRenderer } from '../engine/WorldRenderer';
+import { Block } from './Block';
+import { Car } from './Car';
+import { KeyboardControl, KeyCodeWASD } from './controls/KeyboardControl';
+import { GameStateManager } from './GameStateManager';
+import { CompassRenderable } from './renderables/CompassRenderable';
+import { CurvedGrid } from './renderables/CurvedGrid';
+import { GameStateOverlayRenderable } from './renderables/GameStateOverlayRenderable';
+import { SpeedometerRenderable } from './renderables/SpeedometerRenderable';
+import { TerroristEyesRenderable } from './renderables/TerroristEyesRenderable';
+import { TerroristIndicatorRenderable } from './renderables/TerroristIndicatorRenderable';
+import { Terrorist } from './Terrorist';
 
 export class Game {
     private context = new Context();
