+++ src/main.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+import App from './ui/App';
 import 'tailwindcss/tailwind.css';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
